OvmfPkg/Bhyve/PlatformPei: Fix VS2019 X64 NOOPT build issue
REF: https://bugzilla.tianocore.org/show_bug.cgi?id=3724

Add typecast to fix build error with VS2019 X64 NOOPT converting
a UINT64 value to UINT32 value.

Cc: Rebecca Cran <rebecca@bsdio.com>
Cc: Peter Grehan <grehan@freebsd.org>
Acked-by: Peter Grehan <grehan@freebsd.org>

+++ OvmfPkg/Bhyve/PlatformPei/Platform.c
@@ -191,7 +191,7 @@ MemMapInitialization (
       ASSERT (PciExBarBase <= MAX_UINT32 - SIZE_256MB);
       PciBase = (UINT32)(PciExBarBase + SIZE_256MB);
     } else {
-      PciBase = PcdGet64 (PcdPciMmio32Base);
+      PciBase = (UINT32)PcdGet64 (PcdPciMmio32Base);
       if (PciBase == 0)
         PciBase = (TopOfLowRam < BASE_2GB) ? BASE_2GB : TopOfLowRam;
     }
